requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
cffi==1.14.5
Django==3.2
djangorestframework==3.12.4
pycparser==2.20
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
